feat(TDP-1423): combined shared and web styles
